backend/requirements.txt
fastapi~=0.92.0
lz4~=4.3.2
poetry
msgpack
pydantic~=1.10.5
SQLAlchemy~=2.0.4
starlette~=0.25.0
alembic~=1.9.4
dynaconf~=3.1.11
aiofiles
asyncpg
nats-py
uvicorn
python-multipart
pony